YACL can now be optional(though still is recommended to have!)
BUILD: Probably bumped version again
REFACTOR: Made it so that Explosive Enhancement can be used without YACL. This is to allow for developers using the API to not have to worry about adding YACL as a dependency.
REFACTOR: Added a new "config" package. This is to help with organization(I might add an "integration" or "compat" package later should I add another class for compatibility)
Signed-off-by: Superkat32 <superkat1806@gmail.com>