1.20.2 & version bump (reason for not creating another branch explained below)
BUILD: Bumped version
BUILD: Updated to Minecraft 1.20.2
BUILD: Updated Mod Menu and YACL

Instead of having another branch for 1.20.2, I've decided to keep 1.20.0-1.20.2 in the same branch for now. The hope to is keep one file for all of 1.20, but that might change in the future(should a breaking change be made).

Signed-off-by: Superkat32 <superkat1806@gmail.com>