sadness - 1.21.3 (depression)
Minecraft 1.21.2 removed the data of an explosion's power(which is used to determine an explosion's scale) being sent to the client. While the data can somewhat accurately be calculated using the knockback from an explosion, it simply isn't possible if there is no knockback on the client player. This means that, while on servers, dynamic scaling won't be possible most of the time.

I plan on adding a singleplayer true dynamic scaling option, and will look into server side solutions, but real upsetting honestly.

Signed-off-by: Superkat32 <superkat1806@gmail.com>